Convert PHC and Crypt code to simple serializers/deserializers
Instead of calling back to hash functions, do all the parsing
and serialization in the PHC and Crypt formatters.

They now take a structure, and return it filled with the parameters
or the other way round. This is a little bit simpler and shorter.

The code for the Crypt format was made consistent with the one
for the PHC format, and now accepts arbitrary hash output lengths.

And CryptFormatHasher was rewritten to look similar to
PhcFormatHasher, minus the differences on how the salt is created
and stored.

And error{NoSpaceLeft} is something worth adding to crypto.Error
since it is used a lot here.
A few comments were added and variables were renamed for clarity.